fixed warnings for lack of precision when building promise for 64-bit.

+++ objc-promise/Promise.m
@@ -156,8 +156,8 @@
 
 + (Promise *)or:(NSArray *)promises
 {
-    int count = promises.count;
-    __block int rejectedCount = 0;
+    NSUInteger count = promises.count;
+    __block NSUInteger rejectedCount = 0;
     Deferred *deferred = [[Deferred alloc] init];
     
     // any promise resolves our deferred
@@ -179,8 +179,8 @@
 
 + (Promise *)and:(NSArray *)promises
 {
-    int count = promises.count;
-    __block int resolvedCount = 0;
+    NSUInteger count = promises.count;
+    __block NSUInteger resolvedCount = 0;
     Deferred *deferred = [[Deferred alloc] init];
     
     // any promise resolves our deferred
